feat(Accordion): Add containerStyle prop (#266)
Due to the way that Accordion prop proxies were implemented, it was previously not possible to add a `style` prop that is applied to the Accordion root View component. Added a `containerStyle` prop that Accordion accepts that will be passed as a `style` prop on the Accordion root container View.